migrate system probe to remote agent (only flare)
